requirements.txt
# Telegram Bot
python-telegram-bot==21.10

# Async Database
aiosqlite==0.20.0

# HTTP Client
aiohttp==3.11.11

# Image Processing
Pillow==11.1.0

# Environment Variables
python-dotenv==1.0.1

# Caching
expiringdict==1.2.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
